requirements.txt
line-bot-sdk==1.17.0
Flask==1.1.2
gspread==3.6.0
oauth2client==4.1.3
psycopg2==2.8.6